+++ src/dry_foundation/web/__init__.py
@@ -1,3 +1,4 @@
+from .responses import fetch_json_envelope
 from .static import build_javascript_import_map, static_assets_blueprint
 from .templates import define_basic_template_global_variables
 
@@ -5,4 +6,5 @@ __all__ = [
     "build_javascript_import_map",
     "static_assets_blueprint",
     "define_basic_template_global_variables",
+    "fetch_json_envelope",
 ]