+++ docs/source/CHANGELOG.md
@@ -93,4 +93,5 @@
 - Split database handler `add_match_filter` method into methods for singular and iterable arguments (`add_membership_filter`)
 - Add a `ModelView` object to guard against database alterations for views
 - Provide a `DynamicFieldList` form field object
+- Add the `fetch_json_envelope` decorator for wrapping routes returning raw HTML to return JSON
 - Bundle a minimal sample of common JavaScript modules with the package